requirements.txt
faiss_cpu==1.8.0.post1
pandas==2.2.2
sentence_transformers==3.0.1
streamlit==1.36.0
torch==2.3.1
transformers==4.42.4
